List only installed Operators; some refactor
